requirements.txt
streamlit==1.28.1
pandas==2.0.3
plotly==5.15.0
requests==2.31.0
google-generativeai==0.3.2
bertopic[all]==0.14.0
sentence-transformers==2.2.2
huggingface-hub==0.13.4
scipy==1.11.0
scikit-learn==1.3.0
networkx==3.1
pyvis==0.3.2
